create random exam function updated
Modified it's view: 1- made the weight been taken as an input.
2-  made an input for number of required models for the exam

+++ database/migrations/2021_07_18_223727_exam_models_question_table.php
@@ -17,7 +17,7 @@ class ExamModelsQuestionTable extends Migration
             $table->id();
             $table->foreignId('exam_models_id')->constrained('exam_models')->onDelete('cascade')->onUpdate('cascade');
             $table->foreignId('question_id')->constrained()->onDelete('cascade')->onUpdate('cascade');
-            $table->integer('weight')->nullable(false);
+            $table->float('weight')->nullable(false);
             $table->timestamps();
         });
     }
